.requirements.txt
mysql-connector-python==8.4.0
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
